refactor: 잘못된 레이아웃 삭제.
- 단 한 개의 레이아웃으로 통일, <Layout />
